pyproject.toml
[project]
name = "day-85"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.9"
dependencies = [
 "pillow>=12.0.0",
]
